+++ workspaces/intelligent-assistant/plugins/intelligent-assistant-common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @red-hat-developer-hub/backstage-plugin-intelligent-assistant-common
 
+## 5.3.3
+
 ## 5.3.2
 
 ## 5.3.1